Task 0

+++ 0x15-file_io/0-read_textfile.c
@@ -22,7 +22,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	if (fd == -1)
 		return (0);
 	r = read(fd, buf, letters);
-	if (r == -1 || r != letters)
+	if (r == -1 || r != (ssize_t)letters)
 		return (0);
 	w = write(1, buf, r);
 	if (w == -1 || w != r)
